// week-7/gpa-calculator-app/src/app/transcript.interface.ts
/**
 * Title: transcript.interface.ts
 * Author: Professor Krasso
 * Date: 14 January 2021
 * Modified By: Juvenal Gonzalez
 * Description: interface defines the datatypes of a transcript object
 */


//interface defines what member fields are in a ITranscript Object
export interface ITranscript {
  course: string;
  grade: string;
}